+++ src/pas/plugins/ldap/__init__.py
@@ -7,10 +7,7 @@ from pas.plugins.ldap.plugin import manage_addLDAPPluginForm
 from pas.plugins.ldap.plugin import zmidir
 
 import os
-import monkey
-
-# make Flake 8 happy
-monkey
+import monkey  # noqa
 
 
 def initialize(context):