+++ src/pas/plugins/ldap/interfaces.py
@@ -11,3 +11,22 @@ class ICacheSettingsRecordProvider(Interface):
     """cache settings provider, expects to return a record on call
     In future this may be used more generic.
     """
+
+VALUE_NOT_CACHED = dict()
+
+
+class IPluginCacheHandler(Interface):
+    """Handles caching of the node trees used in the PAS Plugin
+    """
+
+    def get():
+        """the cached value or VALUE_NOT_CACHED
+        """
+
+    def set(value):
+        """sets a value in the cache
+        """
+
+    def invalidate():
+        """removes a value from the cache
+        """